% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ds_gng}
\alias{ds_gng}
\title{Example Go/No Go data}
\format{
An object of class \code{data.frame} with 28200 rows and 7 columns.
}
\usage{
ds_gng
}
\description{
The Go/No Go is a task in which participants respond to one set of stimuli, but withhold a response
to another set of stimuli. This particular dataset is from the first session of a study that is 
described in detail in \doi{10.1016/j.tate.2019.102887}{Hedge, Powell, and Sumner (2018)}. 
It was graciously provided by Craig Hedge and can
be obtained from \href{https://osf.io/cwzds}{https://osf.io/cwzds}.
}
\details{
Overview of columns:

\itemize{
  \item block. Block number
  \item trial. Trial number
  \item stim. Stimuli set used in that block
  \item condition. 0 = go, 2 = no go
  \item response. Correct (1) or incorrect (0)
  \item rt. Reaction time (seconds)
  \item participant. Participant ID
}
}
\keyword{datasets}
