% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ds_rapi}
\alias{ds_rapi}
\title{Example 23-item Rutgers Alcohol Problem Inventory (RAPI) data}
\format{An object of class \code{data.frame} with 426 rows and 24 columns.}
\usage{
ds_rapi
}
\description{
The RAPI is a questionnaire which asks how often a participant experienced each of 23 alcohol-related
problems within the last year.
}
\details{
The dataset contains the following columns:
\itemize{
  \item twnr. Identifies participations
  \item V1 to V23. Answers on each of the 23 RAPI items
}

Each item is answered on a four-point scale with the following answer options:
\itemize{
  \item 0 = None
  \item 1 = 1-2 times
  \item 2 = 3-5
  \item 3 = More than 5 times
}
}
\keyword{datasets}
