% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splithalfr.R
\name{makesplit}
\alias{makesplit}
\title{Splits a list, vector, or data frame, into two random halves of similar size}
\usage{
makesplit(set)
}
\arguments{
\item{set}{(list, vector, or data frame) object to split}
}
\value{
(list) List with two elements that contain each of the two splits of set. Lists and vectors are split by elements, while data frames are split by rows
}
\description{
Splits a list, vector, or data frame, into two random halves of similar size
}
\keyword{internal}
