% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splithalfr.R
\name{mean_rel_by_split}
\alias{mean_rel_by_split}
\title{Reliability coefficient averaged over each split. Can be applied to output of sh_apply}
\usage{
mean_rel_by_split(ds, fn_rel)
}
\arguments{
\item{ds}{(data frame) a data frame with columns "split", "score_1", and "score_2"}

\item{fn_rel}{(function) a function that serves as reliability measure}
}
\value{
(numeric) mean of reliabilities
}
\description{
Reliability coefficient averaged over each split. Can be applied to output of sh_apply
}
\keyword{internal}
