% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splithalfr.R
\name{split_score}
\alias{split_score}
\title{Applies an fn_score on split data}
\usage{
split_score(sets = NULL, fn_score = NULL)
}
\arguments{
\item{sets}{(list) list of vectors, data frames, and lists to feed into fn_score. Each element of sets is split in two random halves of similar size}

\item{fn_score}{(function) called with split sets as argument, should return a single value}
}
\value{
(mixed) Data frame with one row and two columns; score_1 and score_2, which contain the result of the fn_score applied to each split
}
\description{
Applies an fn_score on split data
}
\keyword{internal}
