% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splithalfr.R
\docType{package}
\name{splithalfr}
\alias{splithalfr}
\alias{splithalfr-package}
\title{splithalfr: Extensible Bootstrapped Split-half Reliabilities}
\description{
Calculates scores and bootstrapped split-half reliabilities for reaction time (RT) tasks and questionnaires.
The splithalfr can be extended by custom scoring methods for user-provided datasets.
}
\details{
The splithalfr vignettes demonstrate how to write a custom scoring method based on a set of common RT task scoring methods:
using included example datasets:
\itemize{
  \item \code{vignette("vpt_diff_of_means")} Difference of mean RTs for correct responses, after removing RTs below 200 ms and above 520 ms, on Visual Probe Task data (\href{https://www.tandfonline.com/doi/pdf/10.1080/026999399379050}{Mogg & Bradley, 1999})
  \item \code{vignette("aat_double_diff_of_medians")} Double difference of medians for correct responses on Approach Avoidance Task data (\href{https://psycnet.apa.org/record/1986-17383-001}{MacLeod, Andrews, & Tata, 1986})
  \item \code{vignette("iat_dscore_ri")} Improved d-score algorithm for data of an Implicit Association Task that requires a correct response in order to continue to the next trial \href{https://psycnet.apa.org/record/2003-05897-003}{Greenwald, Nosek, & Banaji, 2003})
}
The code included in each vignette also serves as a test of correct functioning of the splithalfr, by comparing the splithalfr
score for a single participant with the same score calculated via Excel. The materials for each test can be found on
\href{https://github.com/tpronk/splithalfr/tree/master/tests}{the splithalfr GitHub repository}
}
