% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ds_sst}
\alias{ds_sst}
\title{Example Stop Signal Task data}
\format{
An object of class \code{data.frame} with 27000 rows and 7 columns.
}
\usage{
ds_sst
}
\description{
The Stop Signal Task is a task in which participants responded whether a stimulus was a square or a circle.
On 25% of trials the participant heard a tone that indicated that the participant should withhold their response.
This particular dataset is from the first session of a study that is 
described in detail in Hedge et al. (2018) <\doi{10.1016/j.tate.2019.102887}>.
It was graciously provided by Craig Hedge and can
be obtained from \href{https://osf.io/cwzds}{https://osf.io/cwzds}.
}
\details{
Overview of columns:

\itemize{
  \item block. Block number
  \item trial. Trial number
  \item ssd. Stop signal delay
  \item condition. 0 = go, 1 = stop
  \item response. Correct (1) or incorrect (0)
  \item rt. Reaction time (milliseconds)
  \item participant. Participant ID
}
}
\keyword{datasets}
