% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splithalfr.R
\name{flanagan_rulon}
\alias{flanagan_rulon}
\title{Flanagon-Rulon reliability}
\usage{
flanagan_rulon(x, y)
}
\arguments{
\item{x}{(vector) a numeric vector}

\item{y}{(vector) a numeric vector with compatible dimensions to x}
}
\value{
(numeric) Flanagon-Rulon reliability
}
\description{
Flanagon-Rulon reliability
}
\keyword{internal}
