% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splithalfr.R
\name{mean_fr_by_split}
\alias{mean_fr_by_split}
\title{Flanagan-Rulon reliability coefficient averaged over each split. Can be applied to output of sh_apply}
\usage{
mean_fr_by_split(ds)
}
\arguments{
\item{ds}{(data frame) a data frame with columns "split", "score_1", and "score_2"}
}
\value{
(numeric) mean Flanagan-Rulon coefficient
}
\description{
Flanagan-Rulon reliability coefficient averaged over each split. Can be applied to output of sh_apply
}
