% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splithalfr.R
\name{spearman_brown}
\alias{spearman_brown}
\title{Spearman-Brown reliability}
\usage{
spearman_brown(x, y)
}
\arguments{
\item{x}{(vector) a numeric vector}

\item{y}{(vector) a numeric vector with compatible dimensions to x}
}
\value{
(numeric) Spearman-Brown reliability
}
\description{
Spearman-Brown reliability
}
\keyword{internal}
